Update C.3.24.0105 - eConsent enable.feature | Refs [4bbakers/redcap_rsvc#103]
This needs a few adjustments but mostly works.

Needs to be discussed because there are some aspects that did not work as expected.  It does not appear that PDF archive is captured when the "Participant Consent" eConsent is inactive.

Also note that I need to add some aliases for checkbox.

RCTF Improvement needed as well.  Needs to have active/inactive enabled/disabled and toggle switch language added.  These are basically facades that sit on top of the actual functionality but necessary so that human reader can read it properly.